#487: Created empty release calypso module
